AI Lighting torches
When player absorb a fire on the torch, it will attract the attention
to patrolling guards, and pick a guard who is closet to the torch. Then
the guard will go to light the torch while keeping detecting the player.

1. I use 2 TargetPoint array to store the unlit target in GameModeBP,
one is UnlitTorches, the other is UnlitTorches_Reserved. When the
player absorb the torch, it will spawn a TargetPoint object on the
torch position as well as storing the TargetPoint to the UnlitTorches
array that is constantly tracked by all patrolling guards.

2.When patrolling guards find that TargetPoint exists in UnlitTorches
array, they will pick one and move to the target(change State to
“Lighting”) while the chosen target will be moved to the
UnlitTorches_Reserved array in order to prevent guards moving to same
the torch.

3.If guards find the player during Lighting State, they will change to
Combat state, returning the TargetPoint in Reserved array to
UnlitTorches array and to let other guards to light the torch.