Create an Inherited BP, Interface, AIAttack
1. Create a BP named “Character_Base” that is inherited by AICharacter
and FirstPersonalCharacter.

2.Create a new Interface named “HealthInterface”. This contains
interfaces that can be used by “Character_Base” so inherited characters
can have the same interfaces and communication. For example, one
interface name “CauseDamage”. AI can attack the character and send this
interface it. If the character has this interface, it will call the
interface(CauseDamage) like a function.

3.AI now can attack and cause the damage to the character, then the
“CanAttack” boolean will become false. “CheckCanAttack” BTTask will
constantly check when the “CanAttack” is false, it will calculate time
and set  the boolean as true every 2 seconds. This allows AIs to attack
every 2 seconds