Level Control, GameInstance, SaveGame
Create a new map named “Dungeon”
When going to next level(map), save the data like player’s attribute
values, and load it to player character again after loading next level.

The GameInstance BP contains actions for the whole game. Its lifetime
is from the start to the end of the game so it will not be destroyed
when changing levels.

The SaveGame BP is used to provide the game saving functions such as
“save game slot” or “load game from slot”.