Inventory system, collect items, Steal items from AIs
1.Item array is an array of ItemStruct that will be used for recoding
belongings. This will be put in Character, AIs and collectable items.

2.Use line trace with channel to raycast different actors that can be
clicked by the player. For example, the guards brings the jail key in
the Jail level. Players can steal the key if they are close enough to
the guards. Then, the key will be added to the character’s items array
as well as removed from the AI’s items array.